pyproject.toml
[tool.poetry]
name = "sm2t-foss4g"
version = "0.1.0"
description = ""
authors = ["christina <christina.ludwig@uni-heidelberg.de>"]
readme = "README.md"
packages = [{include = "sm2t_foss4g"}]

[tool.poetry.dependencies]
python = "~3.10"
requests = "^2.28.2"
openrouteservice = "^2.3.3"
geopandas = "^0.12.2"
setuptools = "^67.5.1"
matplotlib = "^3.7.1"
tqdm = "^4.65.0"
python-dotenv = "^1.0.0"
seaborn = "^0.12.2"


[tool.poetry.group.dev.dependencies]
pytest = "^7.2.2"
pre-commit = "^3.2.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
